Move commands to commands.c
