+++ include/commands.h
@@ -1,0 +1,8 @@
+
+char* concat(const char *s1, const char *s2);
+
+void ls();
+
+void cd(char *loc);
+
+
